+++ src/main/java/saigonwithlove/ivy/intellij/engine/IvyEngineFactory.java
@@ -47,6 +47,13 @@ public class IvyEngineFactory {
           .definition(this.definition)
           .project(this.project)
           .build();
+    } else if (this.version.getMajorVersion() == 6) {
+      return Ivy6Engine.builder()
+          .directory(this.directory)
+          .version(this.version)
+          .definition(this.definition)
+          .project(this.project)
+          .build();
     }
     throw new NoSuchElementException("Could not create Engine with version: " + this.version);
   }