+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## v0.2.7 - unreleased
+
+### Added
+- Setting for deploy Ivy modules when Ivy Engine started.
+
 ### Fixed
 - Unstable deployment on starting Engine.